changed to double precision floats, added lookahead distance clamp

+++ src/reactive/include/gap_follow_node.hpp
@@ -21,10 +21,10 @@ private:
 
     // Least Squares parameters
     int degree = 2;
-    float steering_gain = 1.0;
-    float lookahead_distance = 1.5; // meters
+    double steering_gain = 1.0;
+    double lookahead_distance = 1.5; // meters
 
-    float max_steering_angle = DEG2RAD(90.0f);
+    double max_steering_angle = DEG2RAD(90.0f);
 
     /// @brief Callback invoked each time we find a valid gap from the laser scan.
     /// @param gap_msg Shared pointer to the incoming Gap message.
@@ -42,23 +42,23 @@ private:
     /// @param r radius of coordinate
     /// @param theta angle of coordinate (in radians)
     /// @returns The Cartesian coordinate
-    std::pair<float, float> polar_to_cartesian(float r, float theta);
+    std::pair<double, double> polar_to_cartesian(double r, double theta);
 
     /// @brief Determine coefficients for the best fitting curve, given a set of inputs (X coordinates) and outputs (Y coordinates)
     /// @param x Vector of X coordinates
     /// @param y Vector of Y coordinates
     /// @returns Vector of coefficients for the polynomial, ordered from lowest degree to highest degree
-    Eigen::VectorXf fit_polynomial(const Eigen::VectorXf &x, const Eigen::VectorXf &y);
+    Eigen::VectorXd fit_polynomial(const Eigen::VectorXd &x, const Eigen::VectorXd &y);
 
     /// @brief Given coefficients and an input value, returns the output value. Basically y = f(x).
     /// @param x Input value
     /// @param coefficients Coefficients of the function, ordered from lowest degree to highest degree
     /// @returns The output value 'y'
-    float get_curve_output(float x, Eigen::VectorXf coefficients);
+    double get_curve_output(float x, Eigen::VectorXd coefficients);
 
     /// @brief Simple method to determine steering angle based on coefficients that were calculated
     /// @param coefficients Coefficients of the function, ordered from lowest degree to highest degree
     /// @param target_x How far ahead the car should look on the curve
     /// @returns The angle the car should steer
-    float compute_steering_angle_simple(Eigen::VectorXf coefficients, float target_x);
+    double compute_steering_angle_simple(Eigen::VectorXd coefficients, double target_x);
 };
